requirements.txt
torch>=1.7.0
torchvision>=0.8.0
opencv-python>=4.5.0
numpy>=1.19.0
scipy>=1.5.0
easydict>=1.9
pillow>=8.0.0
PyYAML>=5.3.1
ultralytics
pandas>=1.1.0
tqdm>=4.50.0
requests>=2.25.0
seaborn>=0.11.0
matplotlib>=3.3.0
